QuarkAudio-UniSE/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.11.0
argbind==0.3.9
asttokens==3.0.0
attrs==25.4.0
audioread==3.1.0
beautifulsoup4==4.14.2
bitarray==3.8.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cmake==4.2.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
Cython==3.2.1
decorator==5.2.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
diff-match-patch==20241021
docstring_parser==0.17.0
easydict==1.13
editdistance==0.8.1
einops==0.8.1
executing==2.2.1
fairseq==0.12.2
fastapi==0.121.2
ffmpy==1.0.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1759948064277/work
fire==0.7.1
flatbuffers==25.9.23
flatten-dict==0.4.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
funasr==1.2.7
future==1.0.0
gdown==5.2.0
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1762946764387/work
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface==0.0.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.0.7
idna==3.11
importlib_resources==6.5.2
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiwer==4.0.0
jmespath==0.10.0
joblib==1.5.2
julius==0.2.7
kaldiio==2.18.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.15.2
llvmlite==0.45.1
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1759055168201/work
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
modelscope==1.31.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
msgpack==1.1.2
multidict==6.7.0
networkx==3.5
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.0.6
onnxruntime==1.23.2
orjson==3.11.4
oss2==2.19.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pesq==0.0.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.19.6
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
pynndescent==0.5.13
pyparsing==3.2.5
PySocks==1.7.1
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytorch-lightning==2.5.6
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1758891829620/work
randomname==0.2.1
RapidFuzz==3.14.3
rarfile==4.2
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rich==14.2.0
ruff==0.14.5
sacrebleu==2.5.1
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
sox==1.5.0
soxr==1.0.0
stack-data==0.6.3
starlette==0.49.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6
termcolor==3.2.0
termplotlib==0.3.9
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.5.1
torch-complex==0.4.4
torch-stoi==0.2.3
torchaudio==2.5.1
torchlibrosa==0.1.0
torchmetrics==1.8.2
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.1.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
webvtt-py==0.5.1
Werkzeug==3.1.3
wget==3.2
wheel==0.45.1
yarl==1.22.0
